Extracted flags to a separate file
